Add geometry types Polygons, Path and Shape
