+++ geoviews/__init__.py
@@ -1,4 +1,4 @@
 from .element import (_Element, Feature, Tiles,     # noqa (API import)
                       WMTS, LineContours, FilledContours, Text, Image,
-                      Points)
+                      Points, Path, Polygons, Shape)
 from . import plotting                              # noqa (API import)
